requirements.txt
wandb==0.16.0
tqdm==4.66.1
PyYAML==6.0.1
seaborn==0.13.0
Pillow==10.1.0
transformers==4.35.2
datasets==2.15.0
pydash==7.0.6
black==23.11.0
python-dotenv==1.0.0
huggingface-hub==0.20.1
pydantic==2.6.1
sentence-transformers==2.3.1
